Arrabiata/tests: move utilities tests into witness_utils
